Update timezone.ts
The function uses the moment-timezone library to create two moment objects, one for the user's timezone and another for the other user's timezone. It then formats the times using the format method with the desired format string, which in this case is "h:mm A z" (hours:minutes AM/PM timezone abbreviation).